Combine 2 plugin versions together
