+++ src/main/resources/hudson/plugins/plot/PlotBuilder/help-name.html
@@ -1,0 +1,9 @@
+<div>
+  Help file for fields are discovered through a file name convention. This file is
+  help for the "name" field. You can have <i>arbitrary</i> HTML here. You can write
+  this file as a Jelly script if you need a dynamic content (but if you do so, change
+  the extension to <tt>.jelly</tt>).
+</div>
+<div>
+  Second text field, that I'm trying to display.
+</div>
