+++ src/main/resources/hudson/plugins/plot/PlotProjectAction/index.jelly
@@ -1,0 +1,35 @@
+<!-- 
+/*
+ * Copyright (c) 2008-2009 Yahoo! Inc.  All rights reserved.
+ * The copyrights to the contents of this file are licensed under the MIT License (http://www.opensource.org/licenses/mit-license.php)
+ */
+ -->
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
+    <l:layout title="${it.project.displayName} plots">
+        <st:include it="${it.project}" page="sidepanel.jelly"/>
+        <l:main-panel>
+            <h1>${%Plot Groups}</h1>
+            <!--<f:textbox default="${it.uniqueName}"></f:textbox>-->
+            <j:choose>
+                <j:when test="${it.hasPlots()}">
+                    <p>
+                        ${%This project contains groups of plots.}
+                    </p>
+                    <table>
+                        <j:forEach var="group" items="${it.originalGroups}">
+                            <t:summary icon="graph.gif" href="${it.getUrlGroup(group)}/">
+                                ${group}
+                            </t:summary>
+                         </j:forEach>
+                    </table>
+                </j:when>
+                <j:otherwise>
+                    <p>
+                        ${%This project has no plots yet.}
+                    </p>
+                </j:otherwise>
+            </j:choose>
+        </l:main-panel>
+    </l:layout>
+</j:jelly>
