+++ src/test/java/hudson/plugins/plot/PlotTest.java
@@ -33,6 +33,8 @@ import hudson.model.BuildListener;
 import hudson.model.AbstractBuild;
 import hudson.model.AbstractProject;
 import hudson.model.FreeStyleProject;
+import hudson.plugins.plot.pipeline.PlotAction;
+import hudson.plugins.plot.pipeline.PlotPublisher;
 import hudson.tasks.Builder;
 import hudson.tasks.LogRotator;
